Disable Goma on riscv.
Goma isn't support on this architecture since it uses gcc for
cross-compilation and the builds silently succeed. However RBE gets used
instead as of the upcoming migration, which is destabilizing the bots
with a too high -j option for local compilation. The right solution is
to simply disble remote compilation on these builders since it's not
working anyway.

Bug: b/296994239
Change-Id: I92c1d4ea082c02ebae9f2ba11196954de556902f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/337800
Commit-Queue: Jonas Termansen <sortie@google.com>
Reviewed-by: Alexander Thomas <athom@google.com>