[dart2wasm] Store the backing array in fixed-size list iterator
Store the Wasm array instead of the List to remove one layer of
indirection when accessing the next element.

We can't do the same in growable list iterator as we need to check
whether the list's length has changed, to be able to throw
`ConcurrentModificationError`.

Change-Id: Iac0fd43dc0c0d249973301903dd34c46fff440f4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/337721
Commit-Queue: Ömer Ağacan <omersa@google.com>
Reviewed-by: Aske Simon Christensen <askesc@google.com>