[dart2js] Use indexed (but not cached) reads for source information.
In the emitter phase source informations are attached to nodes but we have made several changes to ensure the majority of nodes are GCed immediately after being printed. This deserializer cache is holding references to the associated source informations so those objects are not being GCed as well.

Locally see a >400MB memory improvement as a result.

Change-Id: I2fb849c7b0a063579f068bea7d22e7533d61ec05
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/337881
Reviewed-by: Mayank Patke <fishythefish@google.com>
Commit-Queue: Nate Biggs <natebiggs@google.com>