[dart2js] Remove serializer `allowNull` pattern to be more null safe friendly.
With the exception of the one noted case, this change just adds an `OrNull` method variant for any write that was using `allowNull` and changes the call site to use the new `OrNull` method.

There are a few cases where we weren't using the `allowNull` parameter at all. In those cases I've removed the parameter without adding an `OrNull` variant.

Change-Id: Ie5eae4d7f2fb7ccaf9594b9c533afdd93bd99198
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/338280
Reviewed-by: Mayank Patke <fishythefish@google.com>