Wolf analysis: split ValidationEventListener.onAddress
For consistency with other analysis stages (and other code that I've
prototyped but not yet sent for review),
`ValidationEventListener.onAddress` is renamed to `onInstruction`.

Also, its behavior is changed so that it is only called prior to an
instruction in the instruction stream; it is not called at the end of
validation. Instead, `onFinished` is called at the end of validation.

`validator_test.dart` is adjusted accordingly.

Change-Id: I8577006983923c685680c63d7822a5d3a5a42ca3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/339340
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>