Update Dart.g
This change introduces actions that, until implemented, serve as documentation for all the locations where whitespace is considered significant. 

See https://github.com/dart-lang/sdk/issues/50776 for context around significant operator whitespace.

See https://github.com/dart-lang/language/blob/main/accepted/3.0/records/feature-specification.md#ambiguity-with-metadata-annotations for context around significant metadata whitespace. 

It seems to me that there are two ambiguities related to metadata and records. One can be observed when whitespace is present in front of the arguments and a record follows that metadatum, and one when there's no whitespace. We need to disambiguate both.

We disambiguate the case where whitespace is present by requiring the arguments to be preceded by no whitespace.

we disambiguate the case where whitespace is not present by requiring the other metadata productions to have trailing whitespace.