Wolf analysis: use `onAnalysis` callbacks in scope analyzer tests.
Rather than forcing each test to implement an `onInstruction` callback
that checks the address and decides what to do, the test can now
install individual callbacks at any point in the instruction stream
using the pseudo-instruction `onAnalyze`. It requires slightly more
test infrastructure, but it makes the test easier to read, and is more
consistent with the technique used in the validator test.

Change-Id: I6a16ecad9f5b731d8e66fc8ea0a72a1dfc0fadbf
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/339224
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>