Parser: fix error reporting when pattern assignment declares a variable without a name.
The analyzer/CFE shared error template framework requires name
substitutions to be non-empty. However, sometimes, due to parser error
recovery, an identifier token can be an empty string.

One of the circumstances in which this occurs is when parsing text
like the following:

    final b = (final g, final ) = 55;

In this example, there is a parse error due to the fact that the user
is trying to declare variables inside a pattern assignment. But one of
those variables doesn't have a name yet, so the parser has created a
synthetic token for it whose name is an empty string.

To avoid a crash, we must supply some name to the error message
template, so the name we supply is `(unnamed)`.

Fixes https://github.com/dart-lang/sdk/issues/54178.

Bug: https://github.com/dart-lang/sdk/issues/54178
Change-Id: Iabf3263ee1f56a89d1a69bcd74a17a296d470b9c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/339661
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>