[dart2wasm] Avoid duplicate checks in dynamic forwarders
In a forwarder for a method, check class IDs of all classes inheriting
the member and run the shape check once.

For example, in a this class hierarchy:

    class A {
      void printName() { print("A"); }
    }

    class B extends A {
      void printName() { print("B"); }
    }

    class C extends A {}

When generating method forwarder for `printName`, we currently generate
a shape check for all classes.

This this CL we now generate two shape checks: one for A and C (which
share the same `printName`), one for B.

Secondly, when class IDs of a target contains a range of consecutive
class IDs, we now generate a range check, instead of a sequence of
checks for every class ID in the range.

Fixes #53880.

Change-Id: Ie7360075c7930262e9e18c7cd125b83fd32b4bfa
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/339301
Commit-Queue: Ömer Ağacan <omersa@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>